+++ src/QueryDescription/index.js
@@ -5,6 +5,7 @@ import { propEq, pipe, prop, uniq, map } from 'rambdax'
 // don't import whole `utils` to keep worker size small
 import partition from '../utils/fp/partition'
 import invariant from '../utils/common/invariant'
+import deepFreeze from '../utils/common/deepFreeze'
 import type { $RE } from '../types'
 
 import { type TableName, type ColumnName, columnName } from '../Schema'
@@ -28,12 +29,12 @@ export type Operator =
   | 'like'
   | 'notLike'
 
-export type ColumnDescription = $RE<{ column: ColumnName }>
+export type ColumnDescription = $RE<{ column: ColumnName, type?: Symbol }>
 export type ComparisonRight =
   | $RE<{ value: Value }>
   | $RE<{ values: NonNullValues }>
   | ColumnDescription
-export type Comparison = $RE<{ operator: Operator, right: ComparisonRight }>
+export type Comparison = $RE<{ operator: Operator, right: ComparisonRight, type?: Symbol }>
 
 export type WhereDescription = $RE<{
   type: 'where',
@@ -54,6 +55,9 @@ export type On = $RE<{
 export type Condition = Where | On
 export type QueryDescription = $RE<{ where: Where[], join: On[] }>
 
+const columnSymbol = Symbol('Q.column')
+const comparisonSymbol = Symbol('QueryComparison')
+
 // Note: These operators are designed to match SQLite semantics
 // to ensure that iOS, Android, web, and Query observation yield exactly the same results
 //
@@ -71,11 +75,19 @@ export type QueryDescription = $RE<{ where: Where[], join: On[] }>
 //   e.g. `null NOT IN (1, 2, 3) == false`
 
 function _valueOrColumn(arg: Value | ColumnDescription): ComparisonRight {
-  if (arg !== null && typeof arg === 'object') {
-    return arg
+  if (arg === null || typeof arg !== 'object') {
+    return { value: arg }
   }
 
-  return { value: arg }
+  if (typeof arg.column === 'string') {
+    invariant(
+      arg.type === columnSymbol,
+      'Invalid { column: } object passed to Watermelon query. You seem to be passing unsanitized user data to Query builder!',
+    )
+    return { column: arg.column }
+  }
+
+  throw new Error(`Invalid value passed to query`)
 }
 
 // Equals (weakly)
@@ -84,7 +96,7 @@ function _valueOrColumn(arg: Value | ColumnDescription): ComparisonRight {
 // - (1 == true) == true
 // - (0 == false) == true
 export function eq(valueOrColumn: Value | ColumnDescription): Comparison {
-  return { operator: 'eq', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'eq', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Not equal (weakly)
@@ -93,53 +105,52 @@ export function eq(valueOrColumn: Value | ColumnDescription): Comparison {
 // - (1 != true) == false
 // - (0 != false) == false
 export function notEq(valueOrColumn: Value | ColumnDescription): Comparison {
-  return { operator: 'notEq', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'notEq', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Greater than (SQLite semantics)
 // Note:
 // - (5 > null) == false
 export function gt(valueOrColumn: NonNullValue | ColumnDescription): Comparison {
-  return { operator: 'gt', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'gt', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Greater than or equal (SQLite semantics)
 // Note:
 // - (5 >= null) == false
 export function gte(valueOrColumn: NonNullValue | ColumnDescription): Comparison {
-  return { operator: 'gte', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'gte', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Greater than (JavaScript semantics)
 // Note:
 // - (5 > null) == true
 export function weakGt(valueOrColumn: NonNullValue | ColumnDescription): Comparison {
-  return { operator: 'weakGt', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'weakGt', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Less than (SQLite semantics)
 // Note:
 // - (null < 5) == false
 export function lt(valueOrColumn: NonNullValue | ColumnDescription): Comparison {
-  return { operator: 'lt', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'lt', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Less than or equal (SQLite semantics)
 // Note:
 // - (null <= 5) == false
 export function lte(valueOrColumn: NonNullValue | ColumnDescription): Comparison {
-  return { operator: 'lte', right: _valueOrColumn(valueOrColumn) }
+  return { operator: 'lte', right: _valueOrColumn(valueOrColumn), type: comparisonSymbol }
 }
 
 // Value in a set (SQLite IN semantics)
 // Note:
 // - `null` in `values` is not allowed!
 export function oneOf(values: NonNullValues): Comparison {
-  if (process.env.NODE_ENV !== 'production') {
-    invariant(Array.isArray(values), `argument passed to oneOf() is not an array`)
-  }
+  invariant(Array.isArray(values), `argument passed to oneOf() is not an array`)
+  Object.freeze(values) // even in production, because it's an easy mistake to make
 
-  return { operator: 'oneOf', right: { values } }
+  return { operator: 'oneOf', right: { values }, type: comparisonSymbol }
 }
 
 // Value not in a set (SQLite NOT IN semantics)
@@ -147,43 +158,55 @@ export function oneOf(values: NonNullValues): Comparison {
 // - `null` in `values` is not allowed!
 // - (null NOT IN (1, 2, 3)) == false
 export function notIn(values: NonNullValues): Comparison {
-  if (process.env.NODE_ENV !== 'production') {
-    invariant(Array.isArray(values), `argument passed to notIn() is not an array`)
-  }
+  invariant(Array.isArray(values), `argument passed to notIn() is not an array`)
+  Object.freeze(values) // even in production, because it's an easy mistake to make
 
-  return { operator: 'notIn', right: { values } }
+  return { operator: 'notIn', right: { values }, type: comparisonSymbol }
 }
 
 // Number is between two numbers (greater than or equal left, and less than or equal right)
 export function between(left: number, right: number): Comparison {
+  invariant(
+    typeof left === 'number' && typeof right === 'number',
+    'Values passed to Q.between() are not numbers',
+  )
   const values: number[] = [left, right]
-  return { operator: 'between', right: { values } }
+  return { operator: 'between', right: { values }, type: comparisonSymbol }
 }
 
 export function like(value: string): Comparison {
-  return { operator: 'like', right: { value } }
+  invariant(typeof value === 'string', 'Value passed to Q.like() is not string')
+  return { operator: 'like', right: { value }, type: comparisonSymbol }
 }
 
 export function notLike(value: string): Comparison {
-  return { operator: 'notLike', right: { value } }
+  invariant(typeof value === 'string', 'Value passed to Q.notLike() is not string')
+  return { operator: 'notLike', right: { value }, type: comparisonSymbol }
 }
 
 const nonLikeSafeRegexp = /[^a-zA-Z0-9]/g
 
 export function sanitizeLikeString(value: string): string {
+  invariant(typeof value === 'string', 'Value passed to Q.sanitizeLikeString() is not string')
   return value.replace(nonLikeSafeRegexp, '_')
 }
 
 export function column(name: ColumnName): ColumnDescription {
-  return { column: name }
+  invariant(typeof name === 'string', 'Name passed to Q.column() is not string')
+  return { column: name, type: columnSymbol }
 }
 
 function _valueOrComparison(arg: Value | Comparison): Comparison {
-  if (arg !== null && typeof arg === 'object') {
-    return arg
+  if (arg === null || typeof arg !== 'object') {
+    return _valueOrComparison(eq(arg))
   }
 
-  return eq(arg)
+  invariant(
+    arg.type === comparisonSymbol,
+    'Invalid Comparison passed to Query builder. You seem to be passing unsanitized user data to Query builder!',
+  )
+  const { operator, right } = arg
+  return { operator, right }
 }
 
 export function where(left: ColumnName, valueOrComparison: Value | Comparison): WhereDescription {
@@ -244,7 +267,7 @@ export function buildQueryDescription(conditions: Condition[]): QueryDescription
 
   const query = { join, where: whereConditions }
   if (process.env.NODE_ENV !== 'production') {
-    Object.freeze(query)
+    deepFreeze(query)
   }
   return query
 }
@@ -257,7 +280,7 @@ export function queryWithoutDeleted(query: QueryDescription): QueryDescription {
     where: [...whereConditions, whereNotDeleted],
   }
   if (process.env.NODE_ENV !== 'production') {
-    Object.freeze(newQuery)
+    deepFreeze(newQuery)
   }
   return newQuery
 }
@@ -283,7 +306,7 @@ const searchForColumnComparisons: any => boolean = value => {
       // NOTE: To be safe against JS edge cases, there should be hasOwnProperty check
       // but this is performance critical so we trust that this is only called with
       // QueryDescription which doesn't need that
-      if (searchForColumnComparisons(value[key])) {
+      if (key !== 'values' && searchForColumnComparisons(value[key])) {
         return true
       }
     }