+++ src/adapters/__tests__/commonTests.js
@@ -1,6 +1,7 @@
 import expect from 'expect'
 import naughtyStrings from 'big-list-of-naughty-strings'
 
+import expectToRejectWithMessage from '../../__tests__/utils/expectToRejectWithMessage'
 import Model from '../../Model'
 import Query from '../../Query'
 import { sanitizedRaw } from '../../RawRecord'
@@ -20,6 +21,7 @@ import {
   MockTask,
   mockProjectRaw,
   projectQuery,
+  modelQuery,
 } from './helpers'
 
 class BadModel extends Model {
@@ -507,19 +509,16 @@ export default () => [
       await adapter.batch([['create', 'tasks', mockTaskRaw({ id: 't1' })]])
       expect(await adapter.query(taskQuery())).toEqual(['t1'])
 
-      await expect(
+      await expectToRejectWithMessage(
         adapter.batch([
           ['create', 'tasks', mockTaskRaw({ id: 't2' })],
-          ['create', 'does_not_exist', mockTaskRaw({ id: 't3' })],
+          ['create', 'tasks', mockTaskRaw({ id: 't2' })], // duplicate
         ]),
-      ).rejects.toMatchObject({
         // TODO: Get rid of the unknown error - fix on Android
-        message: expect.stringMatching(
-          AdapterClass.name === 'SQLiteAdapter'
-            ? /(no such table: does_not_exist|Exception in HostFunction: <unknown>)/
-            : /Cannot read property 'insert' of null/,
-        ),
-      })
+        AdapterClass.name === 'SQLiteAdapter'
+          ? /(UNIQUE constraint failed: tasks.id|Exception in HostFunction: <unknown>)/
+          : /Duplicate key for property id: t2/,
+      )
       if (AdapterClass.name !== 'LokiJSAdapter') {
         // Regrettably, Loki is not transactional
         expect(await adapter.query(taskQuery())).toEqual(['t1'])
@@ -675,6 +674,79 @@ export default () => [
 
       // deleting already undefined is safe
       await adapter.removeLocal('nonexisting')
+    },
+  ],
+  [
+    'supports naughty strings in LocalStorage',
+    async (adapter, AdapterClass, extraAdapterOptions) => {
+      const usePartialTestBecauseBuggyLoki = AdapterClass.name === 'LokiJSAdapter'
+      if (usePartialTestBecauseBuggyLoki) {
+        // FIXME: https://github.com/techfort/LokiJS/issues/839
+        console.warn('buggy skipping LocalStorage tests') // eslint-disable-line no-console
+      }
+
+      // eslint-disable-next-line no-restricted-syntax
+      for (const string of naughtyStrings) {
+        const key = usePartialTestBecauseBuggyLoki ? `key${Math.random()}` : string
+        // console.log(string)
+        // KNOWN ISSUE: non-JSI adapter implementation gets confused by this (it's a BOM mark)
+        if (
+          AdapterClass.name === 'SQLiteAdapter' &&
+          !extraAdapterOptions.experimentalUseJSI &&
+          string === '﻿'
+        ) {
+          // eslint-disable-next-line no-await-in-loop
+          await adapter.setLocal(key, string)
+          // eslint-disable-next-line no-await-in-loop
+          expect(await adapter.getLocal(key)).not.toBe(string) // if this fails, it means the issue's been fixed
+        } else {
+          // eslint-disable-next-line no-await-in-loop
+          await adapter.setLocal(key, string)
+          // eslint-disable-next-line no-await-in-loop
+          expect(await adapter.getLocal(key)).toBe(string)
+        }
+      }
+    },
+  ],
+  [
+    'does not fail on (weirdly named) table named that are SQLite keywords',
+    async adapter => {
+      await Promise.all(
+        ['where', 'values', 'set', 'drop', 'update'].map(async tableName => {
+          await adapter.batch([['create', tableName, { id: 'i1' }]])
+          await adapter.batch([['update', tableName, { id: 'i1' }]])
+          await adapter.batch([['markAsDeleted', tableName, 'i1']])
+          await adapter.batch([['create', tableName, { id: 'i2' }]])
+          await adapter.find(tableName, 'i2')
+          await adapter.query(modelQuery({ table: tableName }))
+          await adapter.count(modelQuery({ table: tableName }))
+          await adapter.getDeletedRecords(tableName)
+          await adapter.destroyDeletedRecords(tableName, ['i1'])
+          await adapter.batch([['destroyPermanently', tableName, 'i2']])
+          await adapter.getLocal(tableName)
+          await adapter.setLocal(tableName, tableName)
+          await adapter.removeLocal(tableName)
+        }),
+      )
+    },
+  ],
+  [
+    'fails quickly on non-existing table names',
+    async (adapter, AdapterClass) => {
+      const table = 'does-not-exist'
+      const msg = /table name '.*' does not exist/
+      await expectToRejectWithMessage(adapter.find(table, 'i'), msg)
+      await expectToRejectWithMessage(adapter.query(modelQuery({ table })), msg)
+      if (AdapterClass.name === 'SQLiteAdapter') {
+        await expectToRejectWithMessage(adapter.unsafeSqlQuery(table, 'xxx'), msg)
+      }
+      await expectToRejectWithMessage(adapter.count(modelQuery({ table })), msg)
+      await expectToRejectWithMessage(adapter.batch([['create', table, { id: 'i1' }]]), msg)
+      await expectToRejectWithMessage(adapter.batch([['update', table, { id: 'i1' }]]), msg)
+      await expectToRejectWithMessage(adapter.batch([['markAsDeleted', table, 'i1']]), msg)
+      await expectToRejectWithMessage(adapter.batch([['destroyPermanently', table, 'i2']]), msg)
+      await expectToRejectWithMessage(adapter.getDeletedRecords(table), msg)
+      await expectToRejectWithMessage(adapter.destroyDeletedRecords(table, []), msg)
     },
   ],
   [