+++ src/adapters/sqlite/encodeSchema/index.js
@@ -24,7 +24,7 @@ const encodeCreateTable: TableSchema => SQL = ({ name, columns }) => {
 
 const encodeIndex: (ColumnSchema, TableName<any>) => SQL = (column, tableName) =>
   column.isIndexed
-    ? `create index ${tableName}_${column.name} on ${encodeName(tableName)} (${encodeName(
+    ? `create index "${tableName}_${column.name}" on ${encodeName(tableName)} (${encodeName(
         column.name,
       )});`
     : ''
@@ -32,7 +32,7 @@ const encodeIndex: (ColumnSchema, TableName<any>) => SQL = (column, tableName) =
 const encodeTableIndicies: TableSchema => SQL = ({ name: tableName, columns }) =>
   values(columns)
     .map(column => encodeIndex(column, tableName))
-    .concat([`create index ${tableName}__status on ${encodeName(tableName)} ("_status");`])
+    .concat([`create index "${tableName}__status" on ${encodeName(tableName)} ("_status");`])
     .join('')
 
 const encodeTable: TableSchema => SQL = table =>