+++ CHANGELOG.md
@@ -18,7 +18,12 @@ All notable changes to this project will be documented in this file.
 
 ### Changes
 
-- Column and table names starting with `__` and Object property names (e.g. `constructor`) are now forbidden
+- [Hardening] Column and table names starting with `__`, Object property names (e.g. `constructor`), and some reserved keywords are now forbidden
+- [DX] Intializing Database with invalid model classes will now show a helpful error
+- [DX] DatabaseProvider shows a more helpful error if used improperly
+- [Sync] Sync no longer fails if pullChanges returns collections that don't exist on the frontend - shows a warning instead. This is to make building backwards-compatible backends less error-prone
+- [Sync] [Docs] Sync documentation has been rewritten, and is now closer in detail to a formal specification
+- [Hardening] database.collections.get() better validates passed value
 
 ### Fixes